Improve biclique search.
Before, we calculated the set of cubes adjacent to the co-kernels in our
biclique and used that to estimate the upper bound on the count and score of
remaining cubes to add to our biclique. Now, we calculate the cube counts and
scores on a per co-kernel basis, and take the maximum of these.

This gives us a superior upper bound on the maximal value of a biclique
(especially in the "diagonal" case) and also enables us to avoid the cost of
dynamic memory allocation for building the set of adjacent cubes. We have to
perform far more more cube value lookups, but hopefully the cache should make
the effect of this negligible.


Monotone-Parent: 4ff89317e444bc85eb10adbf35345afaeaeea0f0
Monotone-Revision: 1b5440fd467e255e2ab8f371236a7a3de662b3dc