Fix bug in coefficient extraction from sums.
To extract a coefficient from a sum, we use GCD on the numerators and
denominators of the coefficients of the sum's terms. We seed this with
the constant value from the sum. In many cases, this is zero and should
not be used, since it causes denominators to never be extracted.


Monotone-Parent: 7cafc4b2a9c2db138a0afeafdd9598c700ed8515
Monotone-Revision: 6dc25d397e1198cca23fdb570dcab97ac506131c